Need to use floating point numbers if you want floating point math
git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@498 ed5b42cb-e716-0410-a449-f6a68f950b19

+++ src/java/com/threerings/media/timer/CalibratingTimer.java
@@ -113,7 +113,7 @@ public abstract class CalibratingTimer
         long currentTimer = current();
         long currentMillis = System.currentTimeMillis();
         long elapsedTimer = currentTimer - _driftTimerStamp;
-        int elapsedMillis = (int)(currentMillis - _driftMilliStamp);
+        float elapsedMillis = currentMillis - _driftMilliStamp;
         float drift = elapsedMillis / (elapsedTimer / _milliDivider);
         if (_debugCalibrate.getValue()) {
             Log.warning("Calibrating [timer=" + elapsedTimer + ", millis=" + elapsedMillis
